requirements.txt
grpcio-tools>=1.43.0
eth-brownie>=1.18.1
web3>=5.23.1
